This used to be revision r2727.

+++ ui/ui-swt/src/main/java/imagej/ui/swt/display/SWTImageDisplay.java
@@ -150,7 +150,7 @@ public class SWTImageDisplay implements Display {
 	}
 
 	@Override
-	public int[] getCurrentPlanePosition() {
+	public long[] getCurrentPlanePosition() {
 		throw new UnsupportedOperationException("TODO");
 	}
